+++ R/rc_input_amendment.R
@@ -9,56 +9,77 @@
 #' The amendments table used as input for carbon modelling requires at minimum data on effective organic matter inputs and related year.
 #' This helper function assists the checking and controlling of amendments properties involved.
 #'
-#' To run this function, the dt requires as input:"P_NAME", "year","month","P_OM","P_HC","p_p2o5", and "P_DOSE"
-#' if dt is NULL, then the amendment input will be prepared using function \link{rc_input_scenario} using scenario 'BAU'
+#' rothc_amendment: amendment table
+#' Includes the columns:
+#' * P_ID (character), ID of the soil amendment product
+#' * P_NAME (character), name of the soil amendment product, optional
+#' * B_C_OF_INPUT (numeric), the organic carbon input from soil amendment product on a field level (kg C/ha)
+#' * P_DOSE (numeric), applied dose of soil amendment product (kg/ha), required if B_C_OF_INPUT is not supplied
+#' * P_C_OF (numeric), organic carbon content of the soil amendment product (g C/kg), required if B_C_OF_INPUT is not supplied
+#' * P_HC (numeric), the humification coefficient of the soil amendment product (fraction)
+#' * P_DATE_FERTILIZATION (date), date of fertilizer application (formatted YYYY-MM-DD)
 #'
 #' @export
 rc_input_amendment <- function(dt = NULL, B_LU_BRP = NULL){
   
 
   # add visual bindings
-  fr_dpm_rpm = P_HC = cin_tot = P_DOSE = P_OM = cin_hum = cin_dpm = P_NAME = p_p2o5 = cin_rpm = NULL
+  P_C_OF = B_C_OF_INPUT = P_DATE_FERTILIZATION = P_DOSE = P_HC = NULL
+  P_ID = P_NAME = cin_dpm = cin_hum = cin_rpm = cin_tot = fr_dpm_rpm = NULL
   
-  # check B_LU_BRP or crop table
-  checkmate::assert_integerish(B_LU_BRP, any.missing = FALSE, null.ok = TRUE, min.len = 1)
-  checkmate::assert_subset(B_LU_BRP, choices = unique(rotsee::rc_crops$crop_code), empty.ok = TRUE)
-  checkmate::assert_data_table(dt,null.ok = TRUE)
-  checkmate::assert_subset(colnames(dt),choices = c("P_NAME", "year","month","P_OM","P_HC","p_p2o5", "P_DOSE"), empty.ok = TRUE)
-  checkmate::assert_true(!is.null(dt) || !is.null(B_LU_BRP))
-  if(!is.null(dt$month)){checkmate::assert_integerish(dt$month, lower = 1, upper = 12, any.missing = TRUE)}
-  checkmate::assert_numeric(dt$p_p2o5, lower = 0, any.missing = FALSE)
-  checkmate::assert_true(all(dt$p_p2o5 >0))
+  # Check amendment table
+  checkmate::assert_data_table(dt, null.ok = FALSE, min.rows = 1)
+  
+    allowed <- c("P_ID","P_NAME","B_C_OF_INPUT","P_DOSE","P_C_OF","P_HC","P_DATE_FERTILIZATION")
+    checkmate::assert_subset(names(dt), choices = allowed, empty.ok = FALSE)
+  
+    checkmate::assert_true("P_DATE_FERTILIZATION" %in% names(dt))
+    checkmate::assert_date(as.Date(dt$P_DATE_FERTILIZATION), any.missing = FALSE)
+    checkmate::assert_true("P_HC" %in% names(dt))
+    checkmate::assert(
+        "B_C_OF_INPUT" %in% names(dt) || all(c("P_DOSE","P_C_OF") %in% names(dt)),
+        msg = "Provide either B_C_OF_INPUT, or both P_DOSE and P_C_OF"
+      )
+  checkmate::assert_true(!anyNA(dt$B_C_OF_INPUT) || (!anyNA(dt$P_DOSE) & !anyNA(dt$P_C_OF)))
+ 
+  
+  # Create copy of data table
+  dt.org <- copy(dt)
+ 
+  # Add year and month of amendment application
+  dt.org[, year := year(P_DATE_FERTILIZATION)]
+  dt.org[, month := month(P_DATE_FERTILIZATION)]
 
-  
-  if(is.null(dt) & !is.null(B_LU_BRP)){
-    
-    rs <- rc_input_scenario(B_LU_BRP = B_LU_BRP, scen = 'BAU')
-    dt.org <- rs$amendment
-  } else {
-    dt.org <- copy(dt)
-  }
-  
-  
   # Set years to 1:x
   dt.org[,year := year - min(year) + 1]
-  
+ 
   # add month = NA when no input given
   if(!'month' %in% colnames(dt.org)){dt.org[,month := NA_real_]}
-  
-  # add dpm-rmp ratio
+ 
+  # add dpm-rpm ratio
   dt.org[,fr_dpm_rpm := fifelse(P_HC < 0.92, -2.174 * P_HC + 2.02, 0)]
   
-  # estimate total Carbon input per crop and year (kg product * % organic matter * C-fraction = kg C / ha)
-  dt.org[, cin_tot := P_DOSE * P_OM * 0.01 * 0.5]
+  # add default ratio if no humification coefficient is supplied
+  dt.org[is.na(fr_dpm_rpm), fr_dpm_rpm := 1.44]
+  
+  
+  # estimate total Carbon input per crop and year (kg C/ha)
+  if(!is.null(dt.org$B_C_OF_INPUT)){
+    # If supplied, copy value
+    dt.org[, cin_tot := B_C_OF_INPUT]
+  }else{
+    # If not supplied calculate from dose (kg/ha) and organic carbon content (g C/kg)
+  dt.org[, cin_tot := P_DOSE * P_C_OF/1000]
+  }
   
   # estimate C input for DPM, RDM and HUM pool
   dt.org[, cin_hum := 0.02 * cin_tot]
   dt.org[, cin_dpm := (1 - 0.02) * cin_tot * fr_dpm_rpm/ (1 + fr_dpm_rpm)]
   dt.org[, cin_rpm := (1 - 0.02) * cin_tot - cin_dpm]
-  
+ 
   # select only relevant columns
-  dt.org <- dt.org[,list(p_name = P_NAME, year, month, cin_tot, cin_hum, cin_dpm, cin_rpm, fr_eoc_p = P_OM * P_HC * 0.5 / p_p2o5)]
-  
+  dt.org <- dt.org[,list(P_ID = P_ID, P_NAME = P_NAME, year, month, cin_tot, cin_hum, cin_dpm, cin_rpm)]
+
   # return
   return(dt.org)
 }